Refactor separate models out of single file
