+++ pokerapp/pokerbotcontroller.py
@@ -25,6 +25,7 @@ class PokerBotController:
             ('ban', 'Ban user.', self._handle_ban),
             ('cards', 'Show your cards.', self._handle_cards),
             ('reset_game', 'Reset game and refund players', self._reset_game),
+            ('top_up', 'Top up your balance.', self._top_up),
         ]
 
         model._bot.set_my_commands(list(map(lambda e: BotCommand('/' + e[0], e[1]), commands)))
@@ -62,6 +63,9 @@ class PokerBotController:
 
     def _reset_game(self, update: Update, context: CallbackContext) -> None:
         self._model.reset_game(update, context)
+
+    def _top_up(self, update: Update, context: CallbackContext) -> None:
+        self._model.top_up(update, context)
 
     def _handle_button_clicked(self, update: Update, context: CallbackContext) -> None:
         actions = {
