+++ pokerapp/walletmanagermodel.py
@@ -1,0 +1,91 @@
+import datetime
+
+import redis
+
+from pokerapp.entities import Wallet, UserId, Money, UserException
+
+DEFAULT_MONEY = 1000
+
+
+class WalletManagerModel(Wallet):
+    def __init__(self, user_id: UserId, kv: redis.Redis):
+        self.user_id = user_id
+        self._kv = kv
+
+        key = self._prefix(self.user_id)
+        if self._kv.get(key) is None:
+            self._kv.set(key, DEFAULT_MONEY)
+
+    @staticmethod
+    def _prefix(id: int, suffix: str = ""):
+        return "pokerbot:" + str(id) + suffix
+
+    def _current_date(self) -> str:
+        return datetime.datetime.utcnow().strftime("%d/%m/%y")
+
+    def _key_daily(self) -> str:
+        return self._prefix(self.user_id, ":daily")
+
+    def has_daily_bonus(self) -> bool:
+        current_date = self._current_date()
+        last_date = self._kv.get(self._key_daily())
+
+        return last_date is not None and \
+            last_date.decode("utf-8") == current_date
+
+    def add_daily(self, amount: Money) -> Money:
+        if self.has_daily_bonus():
+            raise UserException(
+                "You have already received the bonus today\n"
+                f"Your money: {self.value()}$"
+            )
+
+        key = self._prefix(self.user_id)
+        self._kv.set(self._key_daily(), self._current_date())
+
+        return self._kv.incrby(key, amount)
+
+    def inc(self, amount: Money = 0) -> None:
+        """ Increase count of money in the wallet.
+            Decrease authorized money.
+        """
+        wallet = int(self._kv.get(self._prefix(self.user_id)))
+
+        if wallet + amount < 0:
+            raise UserException("not enough money")
+
+        self._kv.incrby(self._prefix(self.user_id), amount)
+
+    def inc_authorized_money(
+            self,
+            game_id: str,
+            amount: Money
+    ) -> None:
+        key_authorized_money = self._prefix(self.user_id, ":" + game_id)
+        self._kv.incrby(key_authorized_money, amount)
+
+    def authorized_money(self, game_id: str) -> Money:
+        key_authorized_money = self._prefix(self.user_id, ":" + game_id)
+        return int(self._kv.get(key_authorized_money) or 0)
+
+    def authorize(self, game_id: str, amount: Money) -> None:
+        """ Decrease count of money. """
+        self.inc_authorized_money(game_id, amount)
+
+        return self.inc(-amount)
+
+    def authorize_all(self, game_id: str) -> Money:
+        """ Decrease all money of player. """
+        money = int(self._kv.get(self._prefix(self.user_id)))
+        self.inc_authorized_money(game_id, money)
+
+        self._kv.set(self._prefix(self.user_id), 0)
+        return money
+
+    def value(self) -> Money:
+        """ Get count of money in the wallet. """
+        return int(self._kv.get(self._prefix(self.user_id)))
+
+    def approve(self, game_id: str) -> None:
+        key_authorized_money = self._prefix(self.user_id, ":" + game_id)
+        self._kv.delete(key_authorized_money)
